// SurgeonPortal.Api/ClientApp/src/app/api/models/continuouscertification/reference-letter-read-only.model.ts
export interface IReferenceLetterReadOnlyModel {
    id: number;
    userId: number;
    hosp: string;
    official: string;
    title: string;
    roleName: string;
    altRoleName: string;
    roleId: number;
    altRoleId: number;
    explain: string;
    email: string;
    phone: string;
    letterSent: string;
    status: number;
}